+++ src/graphics/geometry/hex.rs
@@ -333,17 +333,17 @@ impl TileGrid {
     /// Linear to rectangular coordinates.
     /// Note this is different from original since the `x` axis is not inverted,
     /// see `from_linear_inv()` for the inverted variant.
-    pub fn from_linear(&self, num: i32) -> Point {
-        Point::new(num % self.width, num / self.width)
+    pub fn from_linear(&self, num: u32) -> Point {
+        Point::new(num as i32 % self.width, num as i32 / self.width)
     }
 
     /// Rectangular to linear coordinates.
     /// Note this is different from original since the `x` axis is not inverted,
     /// see `to_linear_inv()` for the inverted variant.
-    pub fn to_linear(&self, p: impl Into<Point>) -> Option<i32> {
+    pub fn to_linear(&self, p: impl Into<Point>) -> Option<u32> {
         let p = p.into();
         if self.is_in_bounds(p) {
-            Some(self.width * p.y + p.x)
+            Some((self.width * p.y + p.x) as u32)
         } else {
             None
         }
@@ -352,11 +352,11 @@ impl TileGrid {
     /// Rectangular to linear coordinates with `x` axis inverted.
     /// This method should be used when converting linears for use in the original assets
     /// (maps, scripts etc).
-    pub fn to_linear_inv(&self, p: impl Into<Point>) -> Option<i32> {
+    pub fn to_linear_inv(&self, p: impl Into<Point>) -> Option<u32> {
         let p = p.into();
         if self.is_in_bounds(p) {
             let x = self.width - 1 - p.x;
-            Some(self.width * p.y + x)
+            Some((self.width * p.y + x) as u32)
         } else {
             None
         }
@@ -365,9 +365,9 @@ impl TileGrid {
     /// Linear to rectangular coordinates with `x` axis inverted.
     /// This method should be used when converting linears for use in the original assets
     /// (maps, scripts etc).
-    pub fn from_linear_inv(&self, num: i32) -> Point {
-        let x = self.width - 1 - num % self.width;
-        let y = num / self.width;
+    pub fn from_linear_inv(&self, num: u32) -> Point {
+        let x = self.width - 1 - num as i32 % self.width;
+        let y = num as i32 / self.width;
         Point::new(x, y)
     }
